Fixes issue with formatting

+++ composeApp/src/commonMain/kotlin/me/androidbox/spendless/dashboard/presentation/screens/DashboardScreen.kt
@@ -371,7 +371,7 @@ fun LargestTransaction(
 
                 Text(
                     textAlign = TextAlign.Center,
-                    text = largestTransaction.amount.formatMoney(
+                    text = largestTransaction.amount.toDouble().formatMoney(
                         currency = preferenceState.currency,
                         expensesFormat = preferenceState.expensesFormat,
                         thousandsSeparator = preferenceState.thousandsSeparator,
